autons for thirdbot
